backend/api_gateway/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
pillow==10.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2

# Essential packages
# pip install django-cors-headers djangorestframework djangorestframework-simplejwt pillow psycopg2-binary python-dotenv